pyproject.toml
[project]
name = "nerv-mesh"
version = "0.1.0"
description = "AI coding agent with multi-model support via LangGraph + LangChain"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.4",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "lark-oapi>=1.4",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
nerv-mesh = "nerv_mesh.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nerv_mesh"]

[tool.hatch.build.targets.wheel.force-include]
"src/nerv_mesh/_defaults" = "nerv_mesh/_defaults"
"src/nerv_mesh/_builtin_skills" = "nerv_mesh/_builtin_skills"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
